requirements.txt
streamlit>=1.28.0
tensorflow>=2.10.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
plotly>=5.14.0
pandas>=1.5.0
pathlib